finalizando a refatoração do código

+++ src/main/java/com/matheus/gerenciadorDeAlunos/backend/admin/service/AdminService.java
@@ -47,9 +47,9 @@ public class AdminService {
 
     @Transactional
     public void deleteById(UUID id) {
-        try {
+        if (id != null) {
             repository.deleteById(id);
-        } catch (IdNotFoundException exception) {
+        } else {
             throw new IdNotFoundException(id);
         }
     }
